Adding a drafts folder
I'm currently writing content for this blog on two machines, so I'd
like to have a clear way to keep the drafts organized.
I've started to explore a couple of different solutions over the past
few days - the first being to work with a seperate branch for each new
post that gets made. Already I'm working on a queue that has at least 3
articles. Drafts are currently getting created using the hugo new
command and getting put in the content folder with a draft tag in the
metadata at the top of the post which is stored in markdown.

Really, it seems like it would be clearer to use the hugo new command
only when I'm ready to publish a finished post. The drafts could
actually be worked on within a completely seperate drafts folder. This
would allow for the creation of a drafts branch. This is probably
overkill, as each draft is a completely seperate file. Unless we are
working on the same file on two different machines, it is unlikely that
we'll need a seperate branch.